==> 0-0_count_the_number_of_rows_and_columns.ipynb <==
# 12-8-2023-Fri 
# to ount the number of rows and columns of a txt file

file_path = "/mnt/home/royseo/fsgroupAug11/bgbridge/output_lh_aparc.txt"

num_rows = 0
num_columns = 0

with open(file_path, 'r') as file:
    for line in file:
        num_rows +=1
        columns = line.strip().split('\t')
        num_columns = max(num_columns, len(columns))
        
print(f"Number of rows: {num_rows}")
print(f"Number of columns: {num_columns}")
