==> 0-1_preprocessing_paths.ipynb <==
#12-6-2023-Wed
# to extract the cortical volumes from aparc.stats file
# as a first step: I extracted the paths of subjects' aparc.stats file

import pandas as pd
import os
current_directory = os.getcwd()
print ("Current directory:", current_directory)
new = "~"
expanded_directory = os.path.expanduser(new) 
os.chdir(expanded_directory)
updated_directory = os.getcwd()
print ("updated directory:", updated_directory)


listsubj = 'fsgroupAug11/subj'
listsubjtoextract = [d for d in os.listdir(listsubj) if os.path.isdir(os.path.join(listsubj, d)) and d.startswith("sub-")] #get a list of all files and directories starting with "sub-"

file_paths = []
for subj in listsubjtoextract:
    placevolume = '/mnt/home/royseo/fsgroupAug11/subj/' + subj + '/stats/lh.aparc.stats'
    #print(placevolume)
    file_paths.append(placevolume)

output_lh_aparc_path = os.path.expanduser('~/fsgroupAug11/bgbridge/lh_aparc_path.txt')
with open(output_lh_aparc_path,'w') as file:
    for path in file_paths:
        file.write(path + '\n')

file_path = os.path.expanduser('~/fsgroupAug11/bgbridge/lh_aparc_path.txt')

try:
    with open(file_path, 'r') as file:
        line_count = sum(1 for line in file)
        print(f"the file has {line_count} lines.")
        preview_content = file.read()
        print(preview_content)
except FileNotFoundError:
    print(f"File not found: {file_path}")
except Exception as e:
    print(f"An error occurred: {e}")
