==> 2-1-txt2csv.ipynb <==
import csv

txt_file = "output.txt"
csv_file = "output.csv"

with open(txt_file,"r") as in_file, open(csv_file, "w", newline = '') as out_file:
    writer = csv.writer(out_file)
    for line in in_file:
        row = line.strip().split()
        writer.writerow(row)
        

print(csv_file)
