==> 2-2-integrating-handedness-info-and-volume.ipynb <==
import csv
import pandas as pd
import re

%%capture
with open('output.csv','r') as csv_file:
    csv_reader = csv.reader(csv_file)
    for row in csv_reader:
          print(row)
with open('handedness.csv', 'r') as csv_file2:
    csv_reader2 = csv.reader(csv_file2)
    for row2 in csv_reader2:
        print(row2)

%%capture
# output.csv: the size of a output.csv file
df_vol = pd.read_csv('output.csv', header = None)
print(f'Number of rows: {len(df_vol)}')
print(f'Number of columns: {len(df_vol.columns)}')

# output.csv: read the first row; subject ID
subject_list_vol = [df_vol.iloc[0][i] for i in df_vol]
print("this is subject_list_vol", subject_list_vol)

# handedness.csv: the size of handedness.csv file
df_hand = pd.read_csv('handedness.csv')
df_hand = df_hand.loc[:, ~df_hand.columns.str.startswith('Unnamed')]
print(f'Number of rows: {len(df_hand)}')
print(f'Number of columns: {len(df_hand.columns)}')

# handedness.csv: read the first row; subject ID
subj_id_hand = df_hand['Study_ID']
subj_id_right_left = df_hand['Handedness']
print("this is subj_id_hand", subj_id_hand)

%%capture
subj_id_hand_index = []
handedness_in_order = []
error = 0
for i in subject_list_vol:
    a = subject_list_vol.index(i)
    print('*'*50)
    print("index of i is: ", a)
    for ii in subj_id_hand.tolist():
        try: 
            result = re.match(i,ii[0:8])
            if result:
                print(result.group())
                index = subj_id_hand.index[subj_id_hand == ii]
                subj_id_hand_index.append(index[0])
                print(f"The index = {index[0]},{a}")
                b = subj_id_hand.tolist().index(ii)
                handedness = subj_id_right_left.tolist()[b]
                print(f"the handedness = {handedness}")
                handedness_in_order.append(handedness)
                print('*'*50)
        except Exception as e:
            print('-' * 50)
            error += 1
            
print(error)


%%capture
df_vol_transposed = df_vol.transpose()
df_vol_transposed.columns = ['Study_ID',
                             'Left-Lateral-Ventricle',
                                'Left-Inf-Lat-Vent',
                                'Left-Cerebellum-White-Matter',
                                'Left-Cerebellum-Cortex',
                                'Left-Thalamus-Proper',
                                'Left-Caudate',
                                'Left-Putamen',
                                'Left-Pallidum',
                                '3rd-Ventricle',
                                '4th-Ventricle',
                                'Brain-Stem',
                                'Left-Hippocampus',
                                'Left-Amygdala',
                                'CSF',
                                'Left-Accumbens-area',
                                'Left-VentralDC',
                                'Left-vessel',
                                'Left-choroid-plexus',
                                'Right-Lateral-Ventricle',
                                'Right-Inf-Lat-Vent',
                                'Right-Cerebellum-White-Matter',
                                'Right-Cerebellum-Cortex',
                                'Right-Thalamus-Proper',
                                'Right-Caudate',
                                'Right-Putamen',
                                'Right-Pallidum',
                                'Right-Hippocampus',
                                'Right-Amygdala',
                                'Right-Accumbens-area',
                                'Right-VentralDC',
                                'Right-vessel',
                                'Right-choroid-plexus',
                                '5th-Ventricle',
                                'WM-hypointensities',
                                'Left-WM-hypointensities',
                                'Right-WM-hypointensities',
                                'non-WM-hypointensities',
                                'Left-non-WM-hypointensities',
                                'Right-non-WM-hypointensities',
                                'Optic-Chiasm',
                                'CC_Posterior',
                                'CC_Mid_Posterior',
                                'CC_Central',
                                'CC_Mid_Anterior',
                                'CC_Anterior' ]
print(df_vol_transposed)
df_hand['Study_ID']=df_hand['Study_ID'].apply(lambda x: x[:8])
#print(df_hand)

%%capture
# merge two dataframes by subject ID
merged_df = pd.merge(df_vol_transposed, df_hand, on = 'Study_ID')
print(merged_df)

#save results as a csv file
merged_df.to_csv('merged.csv', index = False)
