==> 2-3-IntergratingBehavioralTests.ipynb <==
import pandas as pd

#%%capture
#read MBWC data
file = 'MBWCrepository_NPTdatabase_6.3.2021.xlsx'
sheet= "Database"
df = pd.read_excel(file, sheet_name = sheet)
names =  df.columns.tolist()
print(names)
#shorten the subject ID
df['Study_ID'] = df['Study_ID'].astype(str)
df['Study_ID']=df['Study_ID'].apply(lambda x: x[:8])
output_file_path = "NPT.csv"
df.to_csv(output_file_path, index = False)

#%%capture
# open a csv file that merged to df above
file2 = "merged.csv"
df2 = pd.read_csv(file2)
names2 =  df2.columns.tolist()
print(names2)
# merge the two dataframes (vol,handedess + NPTs)
merged_vol_hand_npt = pd.merge(df, df2, on = 'Study_ID')
#print(merged_vol_hand_npt)

#save results as a csv fi
merged_vol_hand_npt.to_csv('merged_vol_hand_npt.csv', index = False)
