==> 4-1-add-dx-lh_aparc.ipynb <==
# 4-1-add-dx.ipynb was modified on 4-24-2024 for lh_aparc file

import csv
import numpy as np
import pandas as pd

#%%capture
df0 = pd.read_csv('~/fsgroupAug11/bgbridge/2-4-outliersbye-lh_aparc.csv')
print(df0.shape)
print(df0.head())

df1 = pd.read_csv('~/fsgroupAug11/bgbridge/dx-syndrome.csv')
print(df1.shape)
print(df1.head())

names = df1.columns.tolist()
print(names)
print(df1['syndrome_v2_v2'])
dx_counts = df1['syndrome_v2_v2'].value_counts()
print(f"\ndx_counts:  {dx_counts} ")
# 1 Mild Cognitive Impairment #202
# 2 Dementia #162
# 3 Subjective Memory Complaint/Normal Cognition #72
# 4 Unknown - defer to record #5

### add dx to the cleaned.csv
df1 = df1.rename(columns = {'Subject_ID': 'Study_ID'})
#print(df1.head())
df1['Study_ID'] = df1['Study_ID'].str.slice(stop = 8)
print(df1['Study_ID'])

data = pd.merge(df1, df0, on = 'Study_ID')
print(data)

data.to_csv('4-1-add-dx-lh_aparc.csv', index = False)

add_dx_lh_aparc= pd.read_csv('4-1-add-dx-lh_aparc.csv')
add_dx_lh_aparc.columns
add_dx_lh_aparc.columns.shape[0] # 51 columns

add_dx_lh_aparc.columns

# sanity check
print(add_dx_lh_aparc['Age'].mean()) # 70.81025641025641
print(add_dx_lh_aparc['bankssts'].mean()) # 1947.0271125908387
