==> 6-1-SVM-Everything-notyet.ipynb <==
from sklearn.datasets import make_classification
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score 
from sklearn.preprocessing import LabelEncoder

import pandas as pd
# import sys
# #print(sys.version)
# print(sys.executable)
# #from skcuda import svm
# from skcuda import svm
# from skcuda import cudart

#1.dataset load
df = pd.read_csv('data.csv')
pd.set_option('max_columns',None)
#print(list(df.columns))
'''
1 Mild Cognitive Impairment
2 Dementia
3 Subjective Memory Complaint/Normal Cognition
4 Unknown - defer to record
'''
# Filter out unwanted rows and define the label column
df = df[(df['syndrome_v2_v2'] != 4) & (df['syndrome_v2_v2'] != 3)] 
label = 'syndrome_v2_v2' #mild cognitive impairment and Dementia

# Split the data into features and labels
cols_of_interest = [ 'Left-Lateral-Ventricle']
                     'Left-Inf-Lat-Vent']
                                'Left-Cerebellum-White-Matter']
                                'Left-Cerebellum-Cortex']
                                'Left-Thalamus-Proper']
                                'Left-Caudate']
                                'Left-Putamen']
                                'Left-Pallidum']
                                '3rd-Ventricle']
                                '4th-Ventricle'
                                'Brain-Stem']
                                'Left-Hippocampus']
                                'Left-Amygdala']
                                'CSF']
                                'Left-Accumbens-area']
                                'Left-VentralDC']
                                'Left-vessel']
                                'Left-choroid-plexus']
                                'Right-Lateral-Ventricle']
                                'Right-Inf-Lat-Vent']
                                'Right-Cerebellum-White-Matter']
                                'Right-Cerebellum-Cortex']
                                'Right-Thalamus-Proper']
                                'Right-Caudate']
                                'Right-Putamen']
                                'Right-Pallidum']
                                'Right-Hippocampus']
                                'Right-Amygdala']
                                'Right-Accumbens-area']
                                'Right-VentralDC']
                                'Right-vessel']
                                'Right-choroid-plexus']
                                '5th-Ventricle']
                                'WM-hypointensities']
                                'Left-WM-hypointensities']
                                'Right-WM-hypointensities']
                                'non-WM-hypointensities']
                                'Left-non-WM-hypointensities']
                                'Right-non-WM-hypointensities']
                                'Optic-Chiasm']
                                'CC_Posterior']
                                'CC_Mid_Posterior']
                                'CC_Central']
                                'CC_Mid_Anterior']
                                'CC_Anterior']
                    #'Animals_raw']
#  'BNT_totalwstim_raw'] 'FAS_total_raw'] 'TOPF_raw']'WTAR_raw'
X = df[cols_of_interest].copy() # Features
y = df[label].copy()


# Encode labels using LabelEncoder
le = LabelEncoder()
y = le.fit_transform(y)

# Combine Xs and y into a single dataframe 
df = pd.concat([X, pd.DataFrame(y, columns = [label])], axis = 1)

# Drop missing values
df.dropna(inplace = True)


# Separate X and y again
X = df[cols_of_interest]
y = df[label]

#print('X:', X)
print('y:', set(y))

#2.train/test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 123)

#3.Create an SVM instance, nonlinear model
obj = SVC(C = 1.0, kernel = 'rbf', gamma = 'scale')
model = obj.fit(X_train, y_train)

obj1 = SVC(C = 1.0, kernel = 'linear', gamma = 'scale')
model1 = obj1.fit(X_train, y_train)

obj2 = SVC(C = 1.0, kernel = 'poly', gamma = 'scale')
model2 = obj1.fit(X_train, y_train)


obj3 = SVC(C = 1.0, kernel = 'sigmoid', gamma = 'scale')
model3 = obj1.fit(X_train, y_train)

obj4 = SVC(C = 1.0, kernel = 'precomputed', gamma = 'scale')
model4 = obj1.fit(X_train, y_train)

#4. model evaluation
y_pred = model.predict(X = X_test)
acc = accuracy_score(y_test, y_pred)
print('accuracy=', acc)

y_pred = model1.predict(X = X_test)
acc = accuracy_score(y_test, y_pred)
print('accuracy=', acc)

y_pred = model2.predict(X = X_test)
acc = accuracy_score(y_test, y_pred)
print('accuracy=', acc)

y_pred = model3.predict(X = X_test)
acc = accuracy_score(y_test, y_pred)
print('accuracy=', acc)

y_pred = model4.predict(X = X_test)
acc = accuracy_score(y_test, y_pred)
print('accuracy=', acc)

#5. Grid Search
from sklearn.model_selection import GridSearchCV

params = {'kernel': ['rbf', 'linear' ,'sigmoid'], #'poly'], #,'sigmoid'],
         'C': [0.01, 0.1, 1.0, 10.0, 100.0],
         'gamma':['scale', 'auto']} 

grid_model = GridSearchCV(model, param_grid = params, cv = 3, verbose = 1,
                         scoring = 'accuracy', n_jobs = -1).fit(X_test,y_test)

print('best score =', grid_model.best_score_)
print('best parameters = ', grid_model.best_params_)

(x_train, y_train), (x_val, y_val)

X = df[cols_of_interest].copy() # Features
y = df[label].copy()


x_train.shape #(404, 13)
x_val.shape # (102,13)
y_val.shape # (102, )
X, y = boston = load_boston(return_X_y=True)
X.shape # (150, 4)
y.shape # (150,)

from sklearn.model_selection import train_test_split # split
from sklearn.preprocessing import minmax_scale # 정규화(0~1) 
from sklearn.metrics import mean_squared_error, r2_score

import tensorflow as tf # ver 2.x
from tensorflow.keras.datasets import boston_housing
from tensorflow.keras import Sequential # model 생성 
from tensorflow.keras.layers import Dense # DNN layer
