==> 9-1-aseg-lh-aparc-merged.ipynb <==
# to mearge lh-aparc-merged and XX on May 7th 2024

import pandas as pd

aseg = pd.read_csv('~/fsgroupAug11/subj/4-1-add-dx-lh_aparc_modified_for_aseg_stats.csv')
lh_aparc = pd.read_csv('~/fsgroupAug11/bgbridge/4-1-add-dx-lh_aparc.csv')

print(aseg.columns)

print(aseg.columns)

!ls

# Check if Study_ID in lh_aprc and aseg are identical and
# Run merge if they are the same
if aseg['Study_ID'].equals(lh_aparc['Study_ID']):
    merged_df = pd.merge(aseg,lh_aparc, on='Study_ID'
    merged_df.to_csv('~/fsgroupAug11/bgbridge/9-1-aseg-lh-aparc-merged.csv')
    #print(merged_df.head(3))
else:
    print("can't merge the two data frame because the two columns aren't the same")

merged = pd.read_csv('~/fsgroupAug11/bgbridge/9-1-aseg-lh-aparc-merged.csv')

# Verify if the merge operation was performed correctly 
# Check the shape
print("Shape of aseg: ", aseg.shape)
print("Shape of lh_aparc: ", lh_aparc.shape)

print("Shape of merged", merged.shape)

# # Verify if the merge operation was performed correctly 
# Check for missing values

missing_values_count = merged.isnull().sum()
for column_name, missing_count in missing_values_count.items():
    print(f"{column_name}': {missing_count}")
